pyproject.toml
[project]
name = "capable-bench"
version = "0.1.0"
description = "Agentic scientific decision-making benchmark harness for translational peptide reasoning."
requires-python = ">=3.10"
dependencies = [
  "openpyxl>=3.1",
  "pandas>=2.0",
  "pyyaml>=6.0",
]

[project.scripts]
capablebench = "capablebench.cli:main"

